Refactor decodeExpression to ast.
