+++ src/rules-generator.ts
@@ -61,35 +61,6 @@ export interface Validator {
   [name: string]: ValidatorValue;
 };
 
-interface OpPriority {
-  rep?: string;
-  p: number;
-}
-
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Operator_Precedence
-var JS_OPS: { [op: string]: OpPriority; } = {
-  'value': { rep: "", p: 16 },
-
-  'neg': { rep: "-", p: 15},
-  '!': { p: 15},
-  '*': { p: 14},
-  '/': { p: 14},
-  '%': { p: 14},
-  '+': { p: 13 },
-  '-': { p: 13 },
-  '<': { p: 11 },
-  '<=': { p: 11 },
-  '>': { p: 11 },
-  '>=': { p: 11 },
-  'in': { p: 11 },
-  '==': { p: 10 },
-  "!=": { p: 10 },
-  '&&': { p: 6 },
-  '||': { p: 5 },
-  '?:': { p: 4 },
-  ',': { p: 0},
-};
-
 var builtinSchemaNames = ['Any', 'Null', 'String', 'Number', 'Boolean', 'Object'];
 // Method names allowed in Bolt files.
 var valueMethods = ['length', 'includes', 'startsWith', 'beginsWith', 'endsWith',
@@ -250,7 +221,7 @@ export class Generator {
     let keyType = <ast.ExpSimpleType> params[0];
     let valueType = <ast.ExpType> params[1];
     if (keyType.type !== 'type' || !this.symbols.isDerivedFrom(keyType, 'String')) {
-      throw new Error(errors.invalidMapKey + "  (" + decodeExpression(keyType) + " does not)");
+      throw new Error(errors.invalidMapKey + "  (" + ast.decodeExpression(keyType) + " does not)");
     }
 
     let validator = <Validator> {};
@@ -284,7 +255,7 @@ export class Generator {
 
   // Ensure we have a definition for a validator for the given schema.
   ensureValidator(type: ast.ExpType): Validator {
-    var key = decodeExpression(type);
+    var key = ast.decodeExpression(type);
     if (!this.validators[key]) {
       this.validators[key] = {'.validate': ast.literal('***TYPE RECURSION***') };
       this.validators[key] = this.createValidator(type);
@@ -430,7 +401,7 @@ export class Generator {
       !this.isCollectionSchema(schema);
 
     if (hasProps && !this.symbols.isDerivedFrom(schema.derivedFrom, 'Object')) {
-      this.fatal(errors.nonObject + " (is " + decodeExpression(schema.derivedFrom) + ")");
+      this.fatal(errors.nonObject + " (is " + ast.decodeExpression(schema.derivedFrom) + ")");
       return {};
     }
 
@@ -596,7 +567,7 @@ export class Generator {
     // Top level expressions should never be to a snapshot reference - should
     // always evaluate to a boolean.
     exp = ast.ensureBoolean(exp);
-    return decodeExpression(exp);
+    return ast.decodeExpression(exp);
   }
 
   /*
@@ -609,7 +580,7 @@ export class Generator {
   : ast.Exp {
     // Wrap real call for debugging.
     let result = this.partialEvalReal(exp, params, functionCalls);
-    // console.log(decodeExpression(exp) + " => " + decodeExpression(result));
+    // console.log(ast.decodeExpression(exp) + " => " + ast.decodeExpression(result));
     return result;
   }
 
@@ -747,7 +718,7 @@ export class Generator {
         var funcName = ast.getMethodName(expCall);
         if (funcName !== '' && !(funcName in this.symbols.schema['String'].methods ||
               util.arrayIncludes(snapshotMethods, funcName))) {
-          this.fatal(errors.undefinedFunction + decodeExpression(expCall.ref));
+          this.fatal(errors.undefinedFunction + ast.decodeExpression(expCall.ref));
         }
       }
 
@@ -791,7 +762,7 @@ export class Generator {
     }
     var exp = <ast.ExpValue> this.partialEval(args[0], params);
     if (exp.type !== 'String' || !/\/.*\//.test(exp.value)) {
-      throw new Error(errors.coercion + decodeExpression(exp) + " => RegExp");
+      throw new Error(errors.coercion + ast.decodeExpression(exp) + " => RegExp");
     }
     return ast.regexp(exp.value);
   }
@@ -868,126 +839,6 @@ export class Generator {
     this.errorCount += 1;
   }
 };
-
-// From an AST, decode as an expression (string).
-export function decodeExpression(exp: ast.Exp, outerPrecedence?: number): string {
-  if (outerPrecedence === undefined) {
-    outerPrecedence = 0;
-  }
-  var innerPrecedence = precedenceOf(exp);
-  var result;
-  switch (exp.type) {
-  case 'Boolean':
-  case 'Number':
-    result = JSON.stringify((<ast.ExpValue> exp).value);
-    break;
-
-  case 'String':
-    result = util.quoteString((<ast.ExpValue> exp).value);
-    break;
-
-  // RegExp assumed to be in correct format.
-  case 'RegExp':
-    result = (<ast.ExpValue> exp).value;
-    break;
-
-  case 'Array':
-    result = '[' + decodeArray((<ast.ExpValue> exp).value) + ']';
-    break;
-
-  case 'Null':
-    result = 'null';
-    break;
-
-  case 'var':
-  case 'literal':
-    result = (<ast.ExpVariable> exp).name;
-    break;
-
-  case 'ref':
-    let expRef = <ast.ExpReference> exp;
-    if (ast.isIdentifierStringExp(expRef.accessor)) {
-      result = decodeExpression(expRef.base) + '.' + (<ast.ExpValue> expRef.accessor).value;
-    } else {
-      result = decodeExpression(expRef.base, innerPrecedence) +
-        '[' + decodeExpression(expRef.accessor) + ']';
-    }
-    break;
-
-  case 'call':
-    let expCall = <ast.ExpCall> exp;
-    result = decodeExpression(expCall.ref) + '(' + decodeArray(expCall.args) + ')';
-    break;
-
-  case 'builtin':
-    result = decodeExpression(exp);
-    break;
-
-  case 'op':
-    let expOp = <ast.ExpOp> exp;
-    var rep = JS_OPS[expOp.op].rep === undefined ? expOp.op : JS_OPS[expOp.op].rep;
-    if (expOp.args.length === 1) {
-      result = rep + decodeExpression(expOp.args[0], innerPrecedence);
-    } else if (expOp.args.length === 2) {
-      result =
-        decodeExpression(expOp.args[0], innerPrecedence) +
-        ' ' + rep + ' ' +
-        // All ops are left associative - so nudge the innerPrecendence
-        // down on the right hand side to force () for right-associating
-        // operations.
-        decodeExpression(expOp.args[1], innerPrecedence + 1);
-    } else if (expOp.args.length === 3) {
-      result =
-        decodeExpression(expOp.args[0], innerPrecedence) + ' ? ' +
-        decodeExpression(expOp.args[1], innerPrecedence) + ' : ' +
-        decodeExpression(expOp.args[2], innerPrecedence);
-    }
-    break;
-
-  case 'type':
-    result = (<ast.ExpSimpleType> exp).name;
-    break;
-
-  case 'union':
-    result = (<ast.ExpUnionType> exp).types.map(decodeExpression).join(' | ');
-    break;
-
-  case 'generic':
-    let genericType = <ast.ExpGenericType> exp;
-    return genericType.name + '<' + decodeArray(genericType.params) + '>';
-
-  default:
-    result = "***UNKNOWN TYPE*** (" + exp.type + ")";
-    break;
-  }
-
-  if (innerPrecedence < outerPrecedence) {
-    result = '(' + result + ')';
-  }
-
-  return result;
-}
-
-function decodeArray(args: ast.Exp[]): string {
-  return args.map(decodeExpression).join(', ');
-}
-
-function precedenceOf(exp: ast.Exp): number {
-  switch (exp.type) {
-  case 'op':
-    return JS_OPS[(<ast.ExpOp> exp).op].p;
-
-  // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Operator_Precedence
-  // lists call as 17 and ref as 18 - but how could they be anything other than left to right?
-  // http://www.scriptingmaster.com/javascript/operator-precedence.asp - agrees.
-  case 'call':
-    return 18;
-  case 'ref':
-    return 18;
-  default:
-    return 19;
-  }
-}
 
 // Merge all .X terms into target.
 export function extendValidator(target: Validator, src: Validator): Validator {
